Renamed StringParser -> StringUtils
