+++ src/autopas/utils/StringUtils.h
